requirements.txt
amqp==2.2.1
billiard==3.5.0.3
coreapi==2.3.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.4
django-celery==3.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-colorfield==0.1.14
django-debug-toolbar==1.8
django-extensions==1.9.0
django-import-export==0.5.1
django-oidc-provider==0.5.2
django-pipeline==1.6.13
django-redis==4.8.0
django-rest-swagger==2.1.2
django-solo==1.1.2
djangorestframework==3.6.4
mysqlclient==1.3.10
djrill==2.1.0
future==0.16.0
geopy==1.11.0
itypes==1.1.0
kombu==4.1.0
oauthlib==2.0.2
openapi-codec==1.3.2
psycopg2==2.7.3.1
pycryptodomex==3.4.7
pyfcm==1.4.2
pyjwkest==1.3.6
PyJWT==1.5.2
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rollbar==0.13.12
simplejson==3.11.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tablib==0.12.0
uritemplate==3.0.0
vine==1.1.4
whitenoise==3.3.0
xmltodict==0.11.0
